requirements.txt
falcon==4.0.2
pymongo==4.6.1
python-dotenv==1.0.0
marshmallow==4.0.0
gunicorn==23.0.0
pytest==8.3.5
pytest-cov==6.1.1
black==25.1.0
flake8==7.2.0
PyJWT==2.8.0
python-jose==3.4.0
passlib==1.7.4 